backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.4.0
python-dotenv==1.0.0
Werkzeug>=3.0.0
joblib>=1.3.2
